+++ CHANGELOG.md
@@ -14,5 +14,7 @@
  - add examples with json queries
  - add travisCI integration
  - minified version
- - auto prepend _embedded if not spcified
+ - auto prepend `_embedded` if not spcified
  - allow also `filtering_or` operator
+ - parse fields and detect operators
+ - detect if operator is correct or not